remove useless type, move cell styles to appropriate component, handle groups w/ empty members`
